Handle dropped kubeconfig_path from kubernetes-cni interface
The `kubernetes-cni` interface is being changed to no longer include the
`kubeconfig_path` value. Ideally, this would create a specific user for
Tigera EE, but due to the inability to properly test changes to this, a
minimal hard-coded replacement is used instead.

Part of [lp:1920034](https://bugs.launchpad.net/charm-calico/+bug/1920034)